SecondMilestone: Code changed in ImageControllerclass to edit/delete image

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -100,15 +100,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
+        User user = (User) session.getAttribute("loggeduser");
         String tags = convertTagsToString(image.getTags());
+        String error="Only the owner of the image can edit the image";
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-       return "images/edit";
-   }
-
+        //changed the code to check if the user is trying to edit the image is same as the one who uploaded the image..
+        // or else throw the error
+        if(image.getUser().getId()!=user.getId()) {
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+        return "images/edit";
+    }
 
 
 
@@ -152,7 +158,17 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        User user = (User) session.getAttribute("loggeduser");
+        Image deletedImage = imageService.getImage(imageId);
+        String error ="owner has rights to delete the image";
+        model.addAttribute("image", deletedImage);
+        //code to check if the same user is deleting the image is same as the one who uploaded it..
+        // if not error out
+        if(deletedImage.getUser().getId()!=user.getId()) {
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
